trunk: sync with irrlicht 604.

+++ include/vector3d.h
@@ -133,14 +133,15 @@ namespace core
 				getDistanceFromSQ(end) < f;
 		}
 
-		//! Normalizes the vector.
+		//! Normalizes the vector. In case of the 0 vector the result
+		//! is still 0, otherwise the length of the vector will be 1.
 		//! Todo: 64 Bit template doesnt work.. need specialized template
 		vector3d<T>& normalize()
 		{
 			T l = X*X + Y*Y + Z*Z;
 			if (l == 0)
 				return *this;
-			l = (T) reciprocal_squareroot ( l );
+			l = (T) reciprocal_squareroot ( (f32)l );
 			X *= l;
 			Y *= l;
 			Z *= l;
